[Glic] Fix ProfileAttributesEntry GlicEligibility computation
With changes to `glic::GlicEnabling::IsEnabledForProfile()` expecting
to get extended account information from the primary account to
properly perform its computation, the caller now needs to make sure
that the refresh tokens are loaded.

Specifically this created a wrong computation value at profile startup
for the GlicEligibility value in `ProfileAttributesEntry` that is
used to decide whether we show the Profile in the Glic version of the
ProfilePicker or not.

Bug: b:399727295
Change-Id: I0f6e1bdc49fb816571369d9666887b63b8414253
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6357272
Commit-Queue: Ryan Sultanem <rsult@google.com>
Reviewed-by: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1433575}