[ProfilePicker][Glic] Fix GlicController returning the picked profile
With changes to `glic::GlicEnabling::IsEnabledForProfile()` expecting
to get extended account information from the primary account to
properly perform its computation, the caller now needs to make sure
that the refresh tokens are loaded.

To properly support this, after selecting a profile in the Glic
version of ProfilePicker, instead of returning it right away, we now
first ensure that the refresh tokens are loaded. Otherwise, Glic
cannot do much with the profile since it needs the refresh tokens
to properly check for eligibility.

Fixed: b:399727295
Change-Id: I1ef6c26662932e94d6caa82da3dbbb3424c41ca3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6354971
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Ryan Sultanem <rsult@google.com>
Cr-Commit-Position: refs/heads/main@{#1433577}