[code cache] Add Tab Search end-to-end test for bundled code cache
This CL adds an end-to-end test to assert that, when the
WebUIBundledCodeCache feature is enabled, we see the successful load
and consumption of bundled code caches in the renderer for Tab
Search loads.

The test asserts the code cache behavior with field-trials explicitly
enabled and disabled to ensure the feature is resilient to
field-trial variations (this is not the default and handling for this
exists specifically for WebUIBundledCodeCache).

The test only needs to assert the number of successful code cache
loads are greater than 0, since the specific number of code cache
loads is subject to change and a single success demonstrates
a successful end-to-end load into the runtime v8 environment.

Bug: 378504039
Change-Id: Ia9703feddc8737c684ba01d26f7da65df51fea89
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6356451
Reviewed-by: Keren Zhu <kerenzhu@chromium.org>
Commit-Queue: Tom Lukaszewicz <tluk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1433615}