Fixing a bug where you can't drag pinned actions into the last slot
Currently there is a bug where if you attempt to drag a pinned action
into the last slot it doesn't work. This is because we aren't updating
to the correct model index when we attempt to perform a drag due to a
bug in the check we perform during drag. Added a regression test for
this scenario to confirm the fix.

Change-Id: Ia4b0dc5ecd99fc22e9b38005008b9589384ff6cd
Fixed: 403998443
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6359693
Reviewed-by: Caroline Rising <corising@chromium.org>
Commit-Queue: Eshwar Stalin <estalin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1433636}