[AriaNotify] Fire ARIA notifications for newer versions of ATK/Linux
This CL adds the ARIA Notify implementation for Linux.

In versions of the ATK library >= 2.50.0, there is a new "notification"
signal type: https://docs.gtk.org/atk/signal.Object.notification.html

There is more information about this signal type and how it will work in
Orca here:
https://gitlab.gnome.org/GNOME/orca/-/blob/main/README-APPLICATION-DEVELOPERS.md#speaking-your-applications-custom-messageannouncement

In ATK versions that support this signal type, we will use it; in older
versions, we "fall back" to the type of signal that is sent for live
regions. A follow-up CL will have this implementation, CL:6349538.

Bug: 326277796
Change-Id: Ia03971d35c7c16e34bf911ccca9cc5d4174405b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5797671
Reviewed-by: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Commit-Queue: Evelynn Kaplan <evelynn.kaplan@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1433666}