[signin] Create an automation flag for browser signin in tests
This CL adds a flag that promotes web sign in to browser sign-in
automatically.
This is conceptually similar to the setting "Sign in to Chrome when you
sign in to other Google services" being "Sign in" for all accounts.

This flag is useful for automated browsers, to sign in to Chrome
without interacting with Chrome UI.

Bug: 404231491
Change-Id: I9643a6207943e6798ba0800f0615497b2a5c0c3d
Cq-Do-Not-Cancel-Tryjobs: true
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6362358
Reviewed-by: Ryan Sultanem <rsult@google.com>
Commit-Queue: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1433731}