[NTP][Enterprise] Respect hidden modules whenever modules are enabled
When NtpCardsVisible enterprise policy is set to true, all cards are
always enabled. Because of this, sending the list of disabled cards to
the frontend was unnecessary.

However, hidden modules still should be
respected, despite policy, since that is to hide modules that are not
displayed for functionality purposes. For example, Microsoft Outlook
card being hidden until Microsoft authentication is complete.

Bug: 402882633, 402882894
Change-Id: I38ab0402c619e7cc422502052c0f102af82f9f7b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6354010
Reviewed-by: Paul Adedeji <pauladedeji@google.com>
Commit-Queue: Riley Tatum <rtatum@google.com>
Cr-Commit-Position: refs/heads/main@{#1433754}