Allow canvasPromiseTest to run without worker coverage
We sometimes have to test features on the main thread only, but this
was not possible with canvasPromiseTest. That's because
canvasPromiseTest required runCanvasTestsInWorker to be executed as a
precondition, but if no tests are actually enabled for worker,
testharness.js would complain that the worker test file doesn't define
any tests.

This can be addressed by enforcing the precondition that
runCanvasTestsInWorker be called only when a worker test is actually
added. On the flip side, runCanvasTestsInWorker now fails if it's
called in a file that doesn't also added a worker test.

Bug: 403282523
Change-Id: I80cee4e924f4bf782e0baf36cf3f991d33ec1854
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6355064
Commit-Queue: Jean-Philippe Gravel <jpgravel@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1433811}