extensions: Build extension_management_internal.h/cc on desktop Android
Move the ToolbarPinMode enum to a new file to break a circular
dependency back on ExtensionManagement (which does not yet build on
desktop Android).

Do the same with InstallationMode, which is more widely used and
requires more (mechanical) changes.

No functional changes, this just adds one of ExtensionManagement's
dependencies to the build.

Bug: 394876083, 404269488, 404271004
Change-Id: I5a6da907f0e146a16bca1557b4c43ba997766e21
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6358147
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Commit-Queue: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1433847}