[WebContents] Fix crash in GetAllWebContents
Ignore RenderViewHosts that belong to WebContentses that are in the
process of being destroyed. This fixes a crash in GetAllWebContents when
called during destruction of a WebContents. One such situation goes a
little like this:

1. A WM_NCDESTROY arrives for a window.
2. The window destroys its WebContents.
3. The WebContents destroys its RenderFrameHost.
4. The RenderFrameHost destroys its DocumentAssociatedData collection.
5. One such piece of data is a WebBluetoothServiceImpl, which results in
   destruction of a BluetoothAdapterWinrt.
6. BluetoothAdapterWinrt releases a reference to an
   IDeviceInformationStatics.

Now, releasing that reference causes the UI thread's STA to send and
receive messages to/from other apartments/processes. This involves
peeking/dispatching so-called "pending, nonqueued messages". There just
happens to be a WM_SETTINGCHANGE message waiting on the thread's message
queue. This gets dispatched to the browser's message window and leads to
BrowserAccessibilityStateImpl::NotifyWebContentsPreferencesChanged,
which needs to notify all WebContents that WebPreferences have changed.
Kablewie.

Fixed: 362646910
Change-Id: Idba5e7140b2963d1c59b6221f88725bd2b2fa716
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6360435
Auto-Submit: Greg Thompson <grt@chromium.org>
Commit-Queue: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1433938}