Fix crash when multiple start time updates occur for a navigation.
Surprisingly, it is possible for a single NavigationRequest to have its
start time updated multiple times in response to separate beforeunload
completions. We don't have easy repro steps for this, so this is a
speculative fix that removes the failing CHECK in favor of handling the
case in a reasonable way (i.e., preserving the actual original
navigation start time in metrics).

The change should only affect metrics in rare cases, and avoid the
crash.

Bug: 402545469, 385170155
Change-Id: Id33c057c8fb4a3effe22e66c75c465a3372d1597
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6363439
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Auto-Submit: Charlie Reis <creis@chromium.org>
Commit-Queue: Rakina Zata Amni <rakina@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1433943}