Refactor out visibility check logic to `WebContentVisibilityManager`
This CL refactors out the visibility check logic from the
`NavigationPredictorKeyedService` to the `WebContentVisibilityManager`.
The visibility logic is only accessed from the
`SearchEnginePreconnector`, hence we refactor out the logic from the
`NavigationPredictorKeyedService` so that we can use them even if the
keyed service is unavailable.

Bug: 394204688
Change-Id: I2549c8c396040df285a0678af71c5107f86c8dbe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6348541
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Commit-Queue: Keita Suzuki <suzukikeita@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1434023}