Add SearchEnginePreconnectorKeyedServiceFactory
This CL adds `SearchEnginePreconnectorKeyedServiceFactory`, which will
run the `SearchEnginePreconnector` as a KeyedService. This is required
to run an experiment on running the preconnect on Off The Record (OTR)
mode.  Currently, we spawn the `SearchEnginePreconnector` from
`NavigationPredictor`. Since we do not want to enable the predictor and
the related functions for OTR, we refactor them out so that the
Preconnector will run independently.

Bug: 394204688
Change-Id: I88f860b6c874a4dd5d35f62adea567c169cd5283
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6331858
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Commit-Queue: Keita Suzuki <suzukikeita@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1434024}