Enable `SearchEnginePreconnectKeyedService` when feature flag is set
This CL enables `SearchEnginePreconnectKeyedService` when the feature
flag `PreconnectInKeyedService` is set. It will run the
`SearchEnginePreconnector` as a dedicated keyed service instead of running
them as a part of the `NavigationPredictor`. This will allow us to run
the preconnector in Incognito mode as well.

Bug: 394204688
Change-Id: I9021f45eca8113587cf0108df4fe3088bbd54efe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6345354
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Commit-Queue: Keita Suzuki <suzukikeita@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1434025}