[Android] Refactor getAddressUiComponents to use @JniType.
This CL simplifies the AutofillProfileBridge::getAddressUiComponents
API to return AutofillAddressEditorUiInfo class instead of operating
on a bunch of lists. AutofillAddressEditorUiInfo class introduced in
this CL contains a list of address UI components and the best language
tag.

The reason why the list of UI components and the best language tag
cannot be returned separately is because of the libaddressinput API
which doesn't allow that.

Bug: 402020076
Change-Id: Ib22fac5da61b9c0b33d2147f0413b7d4fc366200
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6357049
Reviewed-by: Jan Keitel <jkeitel@google.com>
Commit-Queue: Timofey Chudakov <tchudakov@google.com>
Cr-Commit-Position: refs/heads/main@{#1434035}