[Tab Group Sync] Keep tabs missing groups in the bridge
This CL introduces storing tabs missing groups in the sync bridge. This
is required for some edge cases when a tab group header is not received
by the client yet, so no tabs are lost.

This is the first CL to support tabs missing groups to only store such
tabs in the bridge. Resolving them with new groups will be implemented
in follow-up patches.

Bug: 370719750
Change-Id: Ib4997aa08c562ec9d2d15276e4eb554bc2f34dbd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6350555
Reviewed-by: Shakti Sahu <shaktisahu@chromium.org>
Commit-Queue: Rushan Suleymanov <rushans@google.com>
Cr-Commit-Position: refs/heads/main@{#1434061}