Add IsolatedWebAppThrottle in chrome/browser to wait for WebAppProvider
This CL adds a new IsolatedWebAppThrottle in chrome/browser which can be
used to wait for infra on the chrome side. It also adds the first use
case, i.e. waiting for the WebAppProvider to be ready.
Future CLs will add waiting for the key distribution component and for
manifest parsing.

Bug: 404463500
Change-Id: I5788276fe1c7421b8920a086fa45f585293f6b90
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6360835
Reviewed-by: Andrew Rayskiy <greengrape@google.com>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Commit-Queue: Simon Hangl <simonha@google.com>
Cr-Commit-Position: refs/heads/main@{#1434080}