[MediaDevices] Add relaxed cache mode to EnumerateDevices
In some cases, device enumerations causes the OS to fire notifications
about a possible state change in devices. MediaDevicesManager handled
this by issuing a new enumeration to get the new state. The problem is
that this can result in an endless loop of notifications and
enumerations.

This CL addresses this by introducing a relaxed mode for the cache.
If MDM notices that a number of consecutive notifications results
in no change to the state of media devices, it enters a "relaxed mode"
in which it will ignore notifications for some time, enough to break
the notification/enumeration loop.

Enabled only on Windows.


Bug: 377126111

Change-Id: I5010880adff79f200115d49f64273d335c2d19fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6304750
Commit-Queue: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Palak Agarwal <agpalak@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1434091}