Move Sec Fetch header management on redirects to `FollowRedirect`
`URLLoader::OnReceivedRedirect` sets `defer_redirect` inside of
`URLRequest::ReceivedRedirect` to true, this allows clients to
inspect the redirect and decide what to do with it.
Clients either cancel the request or call `FollowRedirect` which
means that it is enough to calculate and modify Sec headers when
the client decides to follow the redirect.

It should be safe to recompute these Sec Fetch headers inside of
`FollowRedirect` because nothing changes them or relies on them between
`OnReceivedRedirect` and `FollowRedirect`.

This update should not cause any change in behavior so existing tests
should pass without updates.

Bug: 382291442
Change-Id: Ic827254703c5a4145e41dc321fc32273e7c1ab8e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6354048
Reviewed-by: Chris Fredrickson <cfredric@chromium.org>
Commit-Queue: Sandor «Alex» Major <sandormajor@chromium.org>
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1434131}