[graphite] Improve freeing memory when idle
GraphiteCacheController is responsible for freeing memory when the GPU
process is idle. There are 2-3 threads with their own instance of this
class and previously each instance operated independently. That means
the GPU process could be busy drawing when idle cleanup is triggered on
a different thread.

Have GraphiteCacheController wait until all instances are idle instead.
This increments on a global atomic id and only once all instances of the
class are idle will the id not change. Also switch from a 5 second delay
to 1 second delay to match the current GrCacheController implementation.
The idle delay will be between 1-2 seconds now.

Finally if ReduceMemoryUsage() returns true that is a signal from dawn
that there is scheduled GPU work that is still running before objects
can be destroyed. In that case, restart the timer to trigger idle
cleanup again.

Bug: 398193014
Change-Id: I979e0ff0e92be423f32e45dd440f3d32327fa53f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6361483
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Commit-Queue: Kyle Charbonneau <kylechar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1434148}