[iOS] Add Build flags around deprecated API
In iOS 17, the traitCollectionDidChange API was deprecated in favor of
using the registerForTraitCollection changes. It's unclear whether the
workaround the BrowserViewController is implementing here is needed with
this new API paradigm. As a result, I added a build flag around the
deprecated API without a supported replacement.

Fixed: 349385175
Change-Id: I45fc1833dc527b6c25c3e0954c40b22811b52ee6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6357999
Commit-Queue: Daniel White <danieltwhite@google.com>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1434159}