[Protected Audience] Fix origin for coordinator keys in KVv2 cache.
TrustedSignalsCacheImpl was passing the bidder/seller origin to the
coordinator key fetcher instead of the trusted scoring signals origin,
which can be cross-origin to the bidder/seller, now-a-days.

This never made it to stable, and since we require CORS, it likely
wouldn't be a huge security issue, but still good to get this fixed
promptly. It also likely would have broken any cross-origin used of
KVv2, once the cache is enabled.

Fixed: 404285921
Change-Id: I625763cd7d0510cc2822e2de496051e962a80218
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6360610
Reviewed-by: Russ Hamilton <behamilton@google.com>
Commit-Queue: mmenke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1434165}