[MaliciousApkDownloadCheck] Implement random sampling
This is part 5 of 6 of relanding code that was originally landed in
crrev.com/c/6309053 but was reverted.

This CL implements random sampling of eligible downloads, controlled by
a FeatureParam that is configurable via fieldtrial. This is separate
from the sampling of allowlisted downloads and the sampling of
unsupported filetypes, which are both supported on desktop platforms
but not Android.

Bug: 397407934
Change-Id: Ib0f07da7b6e4688c66a0b645e160cb0663a66dab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6341010
Commit-Queue: Lily Chen <chlily@chromium.org>
Reviewed-by: Daniel Rubery <drubery@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1434193}