[omnibox] Update page info icon suppression criteria.
In order to ensure that we're showing the proper experience to users
who do NOT have Google set as their default search engine (DSE), this CL
updates the page info icon suppression logic to make sure that the page
info icon (which contains page-related security info) shows up properly
when the user is in the on-focus zero-prefix state on SRP/Web.

For example, before this CL, the page info icon would get suppressed
when the user focused the Omnibox when visiting non-Google SRP (e.g.
Bing), despite Google SRP being the main surface (besides Web pages)
for on-focus ZPS. This CL basically tightens up the relevant check so
that we're only suppressing the page info icon when it makes sense
(i.e. Google is DSE && non-empty Omnibox popup for ZPS on SRP/Web).

Bug: 400758431
Change-Id: I4ae3a9d539e40115a193e34ff2c03bd9196700ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6362197
Commit-Queue: Khalid Peer <khalidpeer@chromium.org>
Reviewed-by: Orin Jaworski <orinj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1434208}