[MaliciousApkDownloadCheck] Implement telemetry-only mode
This is part 6 of 6 of relanding code that was originally landed in
crrev.com/c/6309053 but was reverted.

This CL implements telemetry-only mode, which is only active for ESB
users, and ignores the verdict in the ClientDownloadResponse.

Compared to the original CL, this CL moves the code that implements
ignoring the verdict to CheckClientDownloadRequestBase, instead of
ChromeDownloadManagerDelegate. This makes it more consistent with
existing conditions that require us to ignore the verdict. In addition,
by providing an UNKNOWN verdict to Chrome, this makes telemetry-only
mode verdicts have no effect on the existing generic APK filetype
warnings (as opposed to suppressing them, which the original CL did,
which was incorrect).

Bug: 397407934
Change-Id: I96f93f8bf8ab87ad4f29c5c45d5410f6db71f955
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6340535
Reviewed-by: Daniel Rubery <drubery@chromium.org>
Commit-Queue: Lily Chen <chlily@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1434231}