[Ad Privacy] Fix the `more` button showing when viewing privacy policy
The `more` button will be shown if it hasn't been hidden before opening the privacy policy. Updated the logic to explicitly hide the `more` button and re-renders the buttons when the back button on the privacy policy is clicked

Screenshot:
State before we open the privacy policy link: https://screenshot.googleplex.com/yAcmnVZYBGNy3Yw
More button showing on privacy policy page:
https://screenshot.googleplex.com/7FZYFHvh9KfffrV

Attached is logic to also remove the top margin for each of the existing dialogs. We were seeing our dialog overlap the action bar on a physical test device which does not seem to be the case anymore.

Bug: 392943234
Change-Id: Ie227b04557037a757357469d75bb972c0dbd50ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6306324
Reviewed-by: Jacob Stanley <jacobstanley@google.com>
Reviewed-by: Abe Boujane <boujane@google.com>
Commit-Queue: Alvin Lam <alvingo@google.com>
Cr-Commit-Position: refs/heads/main@{#1434249}