[lensoverlay] Fix the bucket size of DocumentSize metric
Before, the bucket was using the KB function which defined a bug as:
MIN: 1000 - MAX: 500,000 - # of Buckets - 1000

This had two main issues:
  1) The MIN is too high, causing lots of pages to be bucketed in the
     smallest bucket: https://screenshot.googleplex.com/5rAJpMiCLYyCh4d
  2) Each bucket is 1000 wide, so not a ton of granularity at smaller
     numbers

To fix this, we are now using a custom bucket count with a MIN at 1, MAX
at 100,000, and 100 buckets. This should give the correct level of
granularity.

This also means renaming the histogram as per,
http://shortn/_7zCsHRzqOp.

Bug: 404253089
Change-Id: I9ed4cf9e2a91dd6b2737b8da8c189b8325f62065
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6363654
Reviewed-by: Justin Donnelly <jdonnelly@chromium.org>
Auto-Submit: Duncan Mercer <mercerd@google.com>
Commit-Queue: Duncan Mercer <mercerd@google.com>
Cr-Commit-Position: refs/heads/main@{#1434263}