Tagged PDFs: Moves some accessibility structs in preparation for retrieving the structure tree
In https://crrev.com/c/6339009 the entire structure tree
will be retrieved from the PDF and stored in
an accessibility struct, so that it can be used to later modify
the accessibility tree.

This patch simply moves some existing structs to the right
place in the file so that reviewers will easily know what is a move
and what is new code.

Design doc at:
https://docs.google.com/document/d/1ScD93clMA7AtViWINnaQgRTAK3vbpMshQZQXmWkfCyQ/edit?usp=sharing

AX-Relnotes: n/a.
Bug: 40707542
Change-Id: I6b8a4220ab845bdd84952906dacb4687eafdfcdc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6368296
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1434273}